renamed window.h to sfml_window.h
